LPD-75170 refactor: use the new classes and header with the cell suffix
