+++ modules/dxp/apps/ai-hub/ai-hub-rest-impl/src/main/java/com/liferay/ai/hub/rest/internal/resource/v1_0/MessageResourceImpl.java
@@ -68,7 +68,7 @@ public class MessageResourceImpl extends BaseMessageResourceImpl {
 				contextUser.getUserId()
 			).userToken(
 				contextHttpServletRequest.getHeader(
-					"Liferay-AI-Hub-On-Behalf-Of")
+					"Liferay-AI-Hub-Cell-On-Behalf-Of")
 			).build());
 
 		return message;
